+++ src/renderer/material/material.h
@@ -9,5 +9,5 @@ public:
     // reflect / scatter
     virtual Ray reflect(Object::CollisionData& collision, Ray& incomingRay) = 0;
     // get pixel color. outgoingRay is the ray that is returned from reflect(), otherColor is the color that might have originated from a raytrace
-    virtual glm::vec3 color(Object::CollisionData& collision, Ray& outgoingRay, glm::vec3 otherColor = glm::vec3(1.0f)) = 0;
+    virtual glm::dvec3 color(Object::CollisionData& collision, Ray& outgoingRay, glm::dvec3 otherColor = glm::dvec3(1.0)) = 0;
 };
